Move some segment to dialect package and rename ColumnWithJoinOperatorSegment to OuterJoinExpression
