+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/common/util/ColumnExtractor.java
@@ -24,7 +24,7 @@ import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.BetweenE
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.BinaryOperationExpression;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.InExpression;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.simple.ColumnWithJoinOperatorSegment;
+import org.apache.shardingsphere.sql.parser.sql.dialect.segment.oracle.join.OuterJoinExpression;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.predicate.AndPredicate;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.predicate.WhereSegment;
 
@@ -52,11 +52,11 @@ public final class ColumnExtractor {
             if (((BinaryOperationExpression) expression).getRight() instanceof ColumnSegment) {
                 result.add((ColumnSegment) ((BinaryOperationExpression) expression).getRight());
             }
-            if (((BinaryOperationExpression) expression).getLeft() instanceof ColumnWithJoinOperatorSegment) {
-                result.add(((ColumnWithJoinOperatorSegment) ((BinaryOperationExpression) expression).getLeft()).getColumnName());
+            if (((BinaryOperationExpression) expression).getLeft() instanceof OuterJoinExpression) {
+                result.add(((OuterJoinExpression) ((BinaryOperationExpression) expression).getLeft()).getColumnName());
             }
-            if (((BinaryOperationExpression) expression).getRight() instanceof ColumnWithJoinOperatorSegment) {
-                result.add(((ColumnWithJoinOperatorSegment) ((BinaryOperationExpression) expression).getRight()).getColumnName());
+            if (((BinaryOperationExpression) expression).getRight() instanceof OuterJoinExpression) {
+                result.add(((OuterJoinExpression) ((BinaryOperationExpression) expression).getRight()).getColumnName());
             }
         }
         if (expression instanceof InExpression && ((InExpression) expression).getLeft() instanceof ColumnSegment) {
