+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/dialect/segment/oracle/table/MultiTableConditionalIntoElseSegment.java
@@ -15,36 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr;
+package org.apache.shardingsphere.sql.parser.sql.dialect.segment.oracle.table;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.complex.ComplexExpressionSegment;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.item.ProjectionSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.SQLSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.dml.InsertStatement;
+
+import java.util.Collection;
 
 /**
- * Xml table function segment.
+ * Multi table conditional into else segment.
  */
 @RequiredArgsConstructor
 @Getter
-public final class XmlTableFunctionSegment implements ComplexExpressionSegment, ProjectionSegment {
+public final class MultiTableConditionalIntoElseSegment implements SQLSegment {
     
     private final int startIndex;
     
     private final int stopIndex;
     
-    private final String functionName;
-    
-    private final XmlNameSpacesClauseSegment xmlNameSpacesClause;
-    
-    private final String xQueryString;
-    
-    private final XmlTableOptionsSegment xmlTableOption;
-    
-    private final String text;
-    
-    @Override
-    public String getColumnLabel() {
-        return text;
-    }
+    private final Collection<InsertStatement> insertStatements;
 }
