+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/dialect/segment/oracle/xml/XmlTableOptionsSegment.java
@@ -15,37 +15,30 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr;
+package org.apache.shardingsphere.sql.parser.sql.dialect.segment.oracle.xml;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.ExpressionSegment;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.complex.ComplexExpressionSegment;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.item.ProjectionSegment;
 
 import java.util.Collection;
 import java.util.LinkedList;
 
 /**
- * Xml query and exists function segment.
+ * Xml table options segment.
  */
 @RequiredArgsConstructor
 @Getter
-public final class XmlQueryAndExistsFunctionSegment implements ComplexExpressionSegment, ProjectionSegment {
+public final class XmlTableOptionsSegment implements ComplexExpressionSegment {
     
     private final int startIndex;
     
     private final int stopIndex;
     
-    private final String functionName;
-    
-    private final String xQueryString;
-    
     private final Collection<ExpressionSegment> parameters = new LinkedList<>();
     
-    private final String text;
+    private final Collection<XmlTableColumnSegment> xmlTableColumnSegments = new LinkedList<>();
     
-    @Override
-    public String getColumnLabel() {
-        return text;
-    }
+    private final String text;
 }
