+++ test/it/optimizer/src/test/resources/converter/select-join.xml
@@ -45,4 +45,5 @@
     <test-cases sql-case-id="select_natural_right_join" expected-sql="SELECT * FROM &quot;t_order&quot; AS &quot;o&quot; NATURAL RIGHT JOIN &quot;t_order_item&quot; AS &quot;i&quot; WHERE &quot;o&quot;.&quot;order_id&quot; = ?" db-types="PostgreSQL,openGauss" sql-case-types="PLACEHOLDER" />
     <test-cases sql-case-id="select_natural_full_join" expected-sql="SELECT * FROM &quot;t_order&quot; AS &quot;o&quot; NATURAL FULL JOIN &quot;t_order_item&quot; AS &quot;i&quot; WHERE &quot;o&quot;.&quot;order_id&quot; = 1" db-types="PostgreSQL,openGauss" sql-case-types="LITERAL" />
     <test-cases sql-case-id="select_natural_full_join" expected-sql="SELECT * FROM &quot;t_order&quot; AS &quot;o&quot; NATURAL FULL JOIN &quot;t_order_item&quot; AS &quot;i&quot; WHERE &quot;o&quot;.&quot;order_id&quot; = ?" db-types="PostgreSQL,openGauss" sql-case-types="PLACEHOLDER" />
+    <test-cases sql-case-id="select_with_join_operator" expected-sql="SELECT * FROM t_order o , t_order_item i WHERE o.order_id(+) = i.order_id" db-types="Oracle" />
 </sql-node-converter-test-cases>
