+++ test/it/parser/src/main/java/org/apache/shardingsphere/test/it/sql/parser/internal/cases/parser/jaxb/segment/impl/expr/ExpectedOuterJoinExpression.java
@@ -15,30 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr;
+package org.apache.shardingsphere.test.it.sql.parser.internal.cases.parser.jaxb.segment.impl.expr;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.expr.complex.ComplexExpressionSegment;
+import lombok.Setter;
+import org.apache.shardingsphere.test.it.sql.parser.internal.cases.parser.jaxb.segment.AbstractExpectedSQLSegment;
+import org.apache.shardingsphere.test.it.sql.parser.internal.cases.parser.jaxb.segment.impl.column.ExpectedColumn;
+
+import javax.xml.bind.annotation.XmlElement;
 
 /**
- * Xml table column segment.
+ * Expected outer join expression.
  */
-@RequiredArgsConstructor
 @Getter
-public final class XmlTableColumnSegment implements ComplexExpressionSegment {
+@Setter
+public final class ExpectedOuterJoinExpression extends AbstractExpectedSQLSegment implements ExpectedExpressionSegment {
     
-    private final int startIndex;
+    @XmlElement
+    private ExpectedColumn column;
     
-    private final int stopIndex;
+    @XmlElement(name = "join-operator")
+    private String joinOperator;
     
-    private final String columnName;
-    
-    private final String dataType;
-    
-    private final String path;
-    
-    private final ExpressionSegment defaultExpr;
-    
-    private final String text;
 }
